api/requirements.txt
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-expects-json==1.7.0
# Flask-JWT-Extended==4.4.4
# Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
# mysql==0.0.3
mysql-connector-python==8.0.32
# mysqlclient==2.1.1
protobuf==3.20.3
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
# PyMySQL==1.0.2
pyrsistent==0.19.3
python-dotenv==0.21.1
six==1.16.0
# SQLAlchemy==1.4.46
validators==0.20.0
Werkzeug==2.2.2
